requirements.txt
# Core dependencies for Comprehensive Financial Agent
# Use more flexible version constraints for Windows compatibility

# Async and database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Vector database and embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0

# LangChain ecosystem
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# ChromaDB for local vector storage
chromadb>=0.4.0

# Core data science (use flexible versions)
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# OpenAI and API clients
openai>=1.0.0

# Financial data
yfinance>=0.2.0
finnhub-python>=2.4.0

# Web scraping and parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.65.0
pydantic>=2.0.0
typing-extensions>=4.5.0

# Logging and environment
python-dotenv>=1.0.0
colorlog>=6.7.0
rich>=13.0.0

# Optional: PyTorch (install separately if needed)
# torch>=2.0.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
